requirements.txt
django==5.0.3
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==23.5